Initial import of LTSpiceRender
This is a old project from 2013. Back then it was called "scadutil" and
I used it to transform subcircuits into easily usable macromodels.
However, newer versions of LTSpice now apparently do not detect
subdirectories in models anymore for some reason and also injecting the
files directly into the official bunch (as I did previously) does not
seem to work anymore. Hence I need to figure out how to get it back to
work.