Fixed the issue with non-existing files
The program wasn't broken all along. It was shitty stupid piece of shit
Wine, that dumb-as-a-brick software. Which apparently writes files in my
home directory even when I remove all "dosdevices" symlinks to it. And
since LTSpice in the latest versions copies models from the program
directory to the user data directory, I was always working with the
wrong files.

Geeeeze I hate Wine so much. So much.